feat: добавить класс Card
